Handle trailing comma in method invocations
The parser now adds a `TrailingComma` marker to the last element (unlike Groovy not to the wrapping `JRightPadded` but the element itself) when a trailing comma is encountered in the argument list.

This is also the case when there is a trailing lambda after the method invocation (in which case the marker will be added to the second to last argument).

The printer has also been amended to correctly handle the marker.

Fixes: #228