Fix parsing secondary constructor delegation
It is allowed for a class to not declare any primary constructor but instead secondary constructors. These can then delegate using `this` or `super`.

Additionally, this commit changes how constructors with delegation are mapped in the LST. Instead of "abusing" the return type expression (which for Java is always set to `null` for constructors) and there using a `J.NewClass` (i.e. `new`), the constructor will now receive a body with a `J.MethodInvocation`.

Issue: #270