Support annotations on expressions
There is a new `K.AnnotatedExpression` LST type which holds annotations of annotated expressions and is used to wrap expressions, whenever they are annotated in the source code.

This also required removing `K.KReturn#anntations` and because a return in Kotlin is actually an expression, `KReturn` now also implements `Expression`.

Similarly, a `K.KString` is not a statement, just an expression.

Fixes: #267