Function type parameters support
Up until now the function type were in the LST represented using lambdas. This approach falls down, because with function type parameters the type is mandatory and the name is optional, whereas for lambdas it is the other way around. Also, a lambda as a body, whereas a function type has a return type.

The `K.FunctionType` now declares an explicit `K.FunctionType.Parameters parameters` object, a `Space arrow` for the arrow prefix, and a `TypedTree returnType` for the return type.