Map indexed access `get` operator to method invocation
While the Kotlin index access `get` operator is an operator, its syntax is more like a regular method invocation, with the difference that the method name (`get`) is implicit and that the arguments appear inside square brackets rather than regular parentheses. Also, under the hood, this will of course get mapped to a method invocation (which is also mostly true for Kotlin's other operators).

This commit changes the parsing to map the `get` operator to a `J.MethodInvocation` rather than a `K.Binary` and adds an `IndexedAccess` marker to it, so that it can get printed correctly again.

Fixes: #233